Added Files from Furkan + Added Test Database SQL with test data + Number Only Input Fix

+++ src/main/java/de/hhbk/jahresprojekt/database/Repository.java
@@ -80,7 +80,7 @@ public class Repository<T>  implements CrudRepository<T> {
         SessionFactory sessionFactory = HibernateManager.getSessionFactory();
         Session session = sessionFactory.openSession();
         session.beginTransaction();
-        T object = session.get(tClass, (long)id);
+        T object = session.get(tClass, id);
         session.delete(object);
         session.getTransaction().commit();
         session.close();
